feat: PS/2 keyboard driver with console input abstraction (#277)
Closes #275

- Add ConsoleInput: shared ring buffer (4096 byte cap) that any input
  device can push bytes into. console_read() drains it asynchronously.
  Atomic poll + waker registration prevents race conditions.

- PS/2 keyboard driver (kernel-drivers, feature 'ps2-keyboard'):
  Uses pc-keyboard crate for Scancode Set 1 decoding. Pushes decoded
  ASCII to ConsoleInput via callback. Gated by target_arch x86_64.

- Wire all platforms: UART IRQ handlers on x86_64, aarch64, and
  riscv64 push received bytes to ConsoleInput instead of relying on
  BufferedUart + SerialReadFuture.

- Linux stdin reads from ConsoleInput instead of UART directly.

- Add make clippy-all target (checks aarch64 + x86_64 + riscv64).

- Fix pre-existing clippy issues on x86_64 and riscv64.

+++ zCore/zircon-syscall/src/pci.rs
@@ -121,7 +121,7 @@ impl Syscall<'_> {
                     ((end - addr_win.base) & (PCIE_ECAM_BYTES_PER_BUS as u64 - 1)) as usize;
                 let new_bus_end: usize =
                     addr_win.size / PCIE_ECAM_BYTES_PER_BUS + addr_win.bus_start as usize - 1;
-                if new_bus_end as usize >= PCIE_MAX_BUSSES {
+                if new_bus_end >= PCIE_MAX_BUSSES {
                     return Err(ZxError::INVALID_ARGS);
                 }
                 addr_win.bus_end = new_bus_end as u8;
